chore: update torus-bridge configuration for automatic redirection
- Added a redirect in `next.config.mjs` to automatically route users from the root path to the Fast Bridge page.
- Removed the manual redirect logic from the `HomePage` component, simplifying the code and improving maintainability.

+++ justfile
@@ -22,10 +22,7 @@ gen-types name:
 # --ignore-scripts: don't run scripts (eg. preinstall, postinstall)
 # --frozen-lockfile: don't update the lockfile
 install:
-	if not exists "pnpm-lock.yaml":
-	pnpm install --ignore-scripts
-	else:
-	pnpm install --ignore-scripts --frozen-lockfile
+	[ ! -f pnpm-lock.yaml ] && pnpm install --ignore-scripts || pnpm install --ignore-scripts --frozen-lockfile
 
 # List all packages in the workspace
 ls:
